+++ src/components/native/index.ts
@@ -1,0 +1,5 @@
+import { styled } from "../../theme";
+
+const Div = styled("div", {});
+
+export { Div };