+++ src/components/ui/action/Action.tsx
@@ -1,13 +1,12 @@
 import React from "react";
 import type { ReactNode } from "react";
 
-import classnames from "classnames";
-
 import type { ICON } from "../../icons/Icon";
+import type { ThemeCSS } from "../../../theme";
 
 import { Icon } from "../../icons/Icon";
 
-import styles from "./Action.module.scss";
+import { Anchor, Button, StyledAction } from "./Action.styled";
 
 interface LinkProps { // @TODO extract? - same on list item
     href: string;
@@ -15,28 +14,59 @@ interface LinkProps { // @TODO extract? - same on list item
 }
 
 interface Props {
+    /**
+     * If action should be a native link provide target URL as `href`
+     */
     href?: string;
+    /**
+     * If action should be a routed link provide target URL as `to` and `Link` component
+     */
     to?: string;
-    onClick?: () => void;
-    icon?: ICON | Exclude<ReactNode, string>;
+    /**
+     * If action should be a routed link provide target URL as `to` and `Link` component
+     */
     Link?: React.ComponentClass<LinkProps> | React.FC<LinkProps>;
+    /**
+     * Standard onClick handler (with no event)
+     */
+    onClick?: () => void;
+    /**
+     * Icon to be displayed
+     */
+    icon?: ICON | Exclude<ReactNode, string>;
+    /**
+     * Label to be displayed below the icon
+     */
     label?: ReactNode;
+    /**
+     * Custom class name
+     */
     className?: string;
+    /**
+     * Custom CSS
+     */
+    css?: ThemeCSS;
 }
 
+/**
+ * Action is a round-shaped button or a link, usually used at headers/toolbars.
+ *
+ * Its label is displayed below the circular shape.
+ */
 const Action: React.FC<Props> = (props) => {
-    const { icon, label, href, to, Link, ...restProps } = props;
+    const { icon, label, href, to, Link, css, ...restProps } = props;
+    const maybeCss = css ? { css } : {};
 
     let iconElem: ReactNode = icon;
     if (typeof icon === "string") {
         iconElem = <Icon name={icon as ICON} />;
     }
 
-    const labelElem = label ? <div className={styles.label}>{label}</div> : null;
+    const labelElem = label ? <div>{label}</div> : null;
 
     const content = (
         <>
-            <div className={styles.action}>{iconElem}</div>
+            <StyledAction>{iconElem}</StyledAction>
             {labelElem}
         </>
     );
@@ -46,17 +76,21 @@ const Action: React.FC<Props> = (props) => {
             throw new TypeError("`to` prop given without `Link` component");
         }
 
-        return <Link href={to} {...restProps}><a className={classnames(styles.a, props.className)}>{content}</a></Link>;
+        return (
+            <Link href={to} {...restProps}>
+                <Anchor className={props.className} {...maybeCss}>{content}</Anchor>
+            </Link>
+        );
     }
 
     if (href) {
-        return <a href={href} className={classnames(styles.a, props.className)} {...restProps}>{content}</a>;
+        return <Anchor href={href} className={props.className} {...restProps} {...maybeCss}>{content}</Anchor>;
     }
 
     return (
-        <button onClick={props.onClick} className={classnames(styles.button, props.className)}>
+        <Button onClick={props.onClick} className={props.className} {...maybeCss}>
             {content}
-        </button>
+        </Button>
     );
 };